api/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai
pinecone-client
ragas
datasets
langchain-openai
pandas
tiktoken
